+++ gene2phenotype_project/gene2phenotype_app/serializers/__init__.py
@@ -1,0 +1,20 @@
+from .attrib import AttribSerializer, AttribTypeSerializer
+
+from .user import UserSerializer
+
+from .panel import PanelDetailSerializer
+
+from .publication import PublicationSerializer
+
+from .locus import LocusSerializer, LocusGeneSerializer
+
+from .phenotype import PhenotypeSerializer
+
+from .disease import (DiseaseSerializer, DiseaseOntologyTermSerializer,
+                      CreateDiseaseSerializer, DiseaseDetailSerializer,
+                      GeneDiseaseSerializer)
+
+from .locus_genotype_disease import (LocusGenotypeDiseaseSerializer,
+                                     LGDPanelSerializer)
+
+from .curation import CurationDataSerializer
