+++ gene2phenotype_project/gene2phenotype_app/serializers/disease.py
@@ -1,0 +1,244 @@
+from rest_framework import serializers
+import re
+
+from ..models import (Disease, DiseaseOntologyTerm, DiseaseSynonym,
+                      Attrib, LocusGenotypeDisease, OntologyTerm,
+                      Source, GeneDisease)
+
+from ..utils import (clean_string, get_ontology, get_ontology_source)
+
+
+class DiseaseOntologyTermSerializer(serializers.ModelSerializer):
+    accession = serializers.CharField(source="ontology_term.accession")
+    term = serializers.CharField(source="ontology_term.term")
+    description = serializers.CharField(source="ontology_term.description", allow_null=True)
+    source = serializers.CharField(source="ontology_term.source.name")
+
+    class Meta:
+        model = DiseaseOntologyTerm
+        fields = ['accession', 'term', 'description', 'source']
+
+class DiseaseSerializer(serializers.ModelSerializer):
+    name = serializers.CharField()
+    ontology_terms = serializers.SerializerMethodField()
+    synonyms = serializers.SerializerMethodField()
+
+    def get_ontology_terms(self, id):
+        disease_ontologies = DiseaseOntologyTerm.objects.filter(disease=id)
+        return DiseaseOntologyTermSerializer(disease_ontologies, many=True).data
+
+    def get_synonyms(self, id):
+        synonyms = []
+        disease_synonyms = DiseaseSynonym.objects.filter(disease=id)
+        for d_synonym in disease_synonyms:
+            synonyms.append(d_synonym.synonym)
+        return synonyms
+
+    class Meta:
+        model = Disease
+        fields = ['name', 'ontology_terms', 'synonyms']
+
+class DiseaseDetailSerializer(DiseaseSerializer):
+    last_updated = serializers.SerializerMethodField()
+
+    def get_last_updated(self, id):
+        filtered_lgd_list = LocusGenotypeDisease.objects.filter(
+            disease=id,
+            is_reviewed=1,
+            is_deleted=0,
+            date_review__isnull=False
+            ).latest('date_review'
+                     ).date_review
+
+        return filtered_lgd_list.date() if filtered_lgd_list else []
+
+    def records_summary(self, id, user):
+        lgd_list = LocusGenotypeDisease.objects.filter(disease=id, is_deleted=0)
+
+        if user.is_authenticated:
+            lgd_select = lgd_list.select_related('disease', 'genotype', 'confidence'
+                                               ).prefetch_related('lgd_panel', 'panel', 'lgd_variant_gencc_consequence', 'lgd_variant_type', 'lgd_molecular_mechanism', 'g2pstable_id'
+                                                                  ).order_by('-date_review')
+
+        else:
+            lgd_select = lgd_list.select_related('disease', 'genotype', 'confidence'
+                                               ).prefetch_related('lgd_panel', 'panel', 'lgd_variant_gencc_consequence', 'lgd_variant_type', 'lgd_molecular_mechanism', 'g2pstable_id'
+                                                                  ).order_by('-date_review').filter(lgdpanel__panel__is_visible=1)
+
+
+        lgd_objects_list = list(lgd_select.values('disease__name',
+                                                  'lgdpanel__panel__name',
+                                                  'stable_id__stable_id', # to get the stable_id stableID
+                                                  'genotype__value',
+                                                  'confidence__value',
+                                                  'lgdvariantgenccconsequence__variant_consequence__term',
+                                                  'lgdvarianttype__variant_type_ot__term',
+                                                  'lgdmolecularmechanism__mechanism__value'))
+
+        aggregated_data = {}
+        for lgd_obj in lgd_objects_list:
+            if lgd_obj['stable_id__stable_id'] not in aggregated_data.keys():
+                variant_consequences = []
+                variant_types = []
+                molecular_mechanism = []
+                panels = []
+
+                panels.append(lgd_obj['lgdpanel__panel__name'])
+                variant_consequences.append(lgd_obj['lgdvariantgenccconsequence__variant_consequence__term'])
+                if lgd_obj['lgdvarianttype__variant_type_ot__term'] is not None:
+                    variant_types.append(lgd_obj['lgdvarianttype__variant_type_ot__term'])
+                if lgd_obj['lgdmolecularmechanism__mechanism__value'] is not None:
+                    molecular_mechanism.append(lgd_obj['lgdmolecularmechanism__mechanism__value'])
+
+                aggregated_data[lgd_obj['stable_id__stable_id']] = { 'disease':lgd_obj['disease__name'],
+                                                          'genotype':lgd_obj['genotype__value'],
+                                                          'confidence':lgd_obj['confidence__value'],
+                                                          'panels':panels,
+                                                          'variant_consequence':variant_consequences,
+                                                          'variant_type':variant_types,
+                                                          'molecular_mechanism':molecular_mechanism,
+                                                          'stable_id':lgd_obj['stable_id__stable_id'] }
+
+            else:
+                if lgd_obj['lgdpanel__panel__name'] not in aggregated_data[lgd_obj['stable_id__stable_id']]['panels']:
+                    aggregated_data[lgd_obj['stable_id__stable_id']]['panels'].append(lgd_obj['lgdpanel__panel__name'])
+                if lgd_obj['lgdvariantgenccconsequence__variant_consequence__term'] not in aggregated_data[lgd_obj['stable_id__stable_id']]['variant_consequence']:
+                    aggregated_data[lgd_obj['stable_id__stable_id']]['variant_consequence'].append(lgd_obj['lgdvariantgenccconsequence__variant_consequence__term'])
+                if lgd_obj['lgdvarianttype__variant_type_ot__term'] not in aggregated_data[lgd_obj['stable_id__stable_id']]['variant_type'] and lgd_obj['lgdvarianttype__variant_type_ot__term'] is not None:
+                    aggregated_data[lgd_obj['stable_id__stable_id']]['variant_type'].append(lgd_obj['lgdvarianttype__variant_type_ot__term'])
+                if lgd_obj['lgdmolecularmechanism__mechanism__value'] not in aggregated_data[lgd_obj['stable_id__stable_id']]['molecular_mechanism'] and lgd_obj['lgdmolecularmechanism__mechanism__value'] is not None:
+                    aggregated_data[lgd_obj['stable_id__stable_id']]['molecular_mechanism'].append(lgd_obj['lgdmolecularmechanism__mechanism__value'])
+
+        return aggregated_data.values()
+
+    class Meta:
+        model = Disease
+        fields = DiseaseSerializer.Meta.fields + ['last_updated']
+
+class CreateDiseaseSerializer(serializers.ModelSerializer):
+    ontology_terms = DiseaseOntologyTermSerializer(many=True, required=False)
+
+    # Add synonyms
+
+    def create(self, validated_data):
+        disease_name = validated_data.get('name')
+        ontologies_list = validated_data.get('ontology_terms')
+
+        disease_obj = None
+
+        # Clean disease name
+        cleaned_input_disease_name = clean_string(str(disease_name))
+        # Check if name already exists
+        all_disease_names = Disease.objects.all()
+        for disease_db in all_disease_names:
+            cleaned_db_disease_name = clean_string(str(disease_db.name))
+            if cleaned_db_disease_name == cleaned_input_disease_name:
+                disease_obj = disease_db
+        all_disease_synonyms = DiseaseSynonym.objects.all()
+        for disease_synonym in all_disease_synonyms:
+            cleaned_db_disease_syn = clean_string(str(disease_synonym.synonym))
+            if cleaned_db_disease_syn == cleaned_input_disease_name:
+                disease_obj = disease_synonym.disease
+
+        if disease_obj is None:
+            # TODO: give disease suggestions
+
+            disease_obj = Disease.objects.create(
+                name = disease_name
+            )
+
+        # Check if ontology is in db
+        # The disease ontology is saved in the db as attrib type 'disease'
+        for ontology in ontologies_list:
+            ontology_accession = ontology['ontology_term']['accession']
+            ontology_term = ontology['ontology_term']['term']
+            ontology_desc = ontology['ontology_term']['description']
+            disease_ontology_obj = None
+
+            if ontology_accession is not None and ontology_term is not None:
+                try:
+                    ontology_obj = OntologyTerm.objects.get(accession=ontology_accession)
+
+                except OntologyTerm.DoesNotExist:
+                    # Check if ontology is from OMIM or Mondo
+                    source = get_ontology_source(ontology_accession)
+
+                    if source is None:
+                        raise serializers.ValidationError({
+                            "message": f"Invalid ID '{ontology_accession}' please input a valid ID from OMIM or Mondo"
+                            })
+
+                    elif source == "Mondo":
+                        # Check if ontology accession is valid
+                        mondo_disease = get_ontology(ontology_accession, source)
+                        if mondo_disease is None:
+                            raise serializers.ValidationError({"message": "Invalid Mondo ID",
+                                                                   "Please check ID": ontology_accession})
+                        elif mondo_disease == "query failed":
+                            raise serializers.ValidationError({"message": f"Cannot query Mondo ID {ontology_accession}"})
+
+                    # Replace '_' from mondo ID
+                    ontology_accession = re.sub(r'\_', ':', ontology_accession)
+                    ontology_term = re.sub(r'\_', ':', ontology_term)
+                    # Insert ontology
+                    if ontology_desc is None and len(mondo_disease['description']) > 0:
+                        ontology_desc = mondo_disease['description'][0]
+
+                    elif source == "OMIM":
+                        omim_disease = get_ontology(ontology_accession, source)
+                        # TODO: check if we can use the OMIM API in the future
+                        if omim_disease == "query failed":
+                            raise serializers.ValidationError({"message": f"Cannot query OMIM ID {ontology_accession}"})
+
+                        if ontology_desc is None and omim_disease is not None and len(omim_disease['description']) > 0:
+                            ontology_desc = omim_disease['description'][0]
+
+                    source = Source.objects.get(name=source)
+                    # Get attrib 'disease'
+                    attrib_disease = Attrib.objects.get(
+                        value = "disease",
+                        type__code = "ontology_term_group"
+                    )
+
+                    ontology_obj = OntologyTerm.objects.create(
+                                accession = ontology_accession,
+                                term = ontology_term,
+                                description = ontology_desc,
+                                source = source,
+                                group_type = attrib_disease
+                    )
+
+                attrib = Attrib.objects.get(
+                    value="Data source",
+                    type__code = "ontology_mapping"
+                )
+
+                try:
+                    # Check if disease-ontology is stored in G2P
+                    disease_ontology_obj = DiseaseOntologyTerm.objects.get(
+                        disease = disease_obj,
+                        ontology_term = ontology_obj,
+                        mapped_by_attrib = attrib,
+                    )
+                except DiseaseOntologyTerm.DoesNotExist:
+                    # Insert disease-ontology
+                    disease_ontology_obj = DiseaseOntologyTerm.objects.create(
+                        disease = disease_obj,
+                        ontology_term = ontology_obj,
+                        mapped_by_attrib = attrib,
+                    )
+
+        return disease_obj
+
+    class Meta:
+        model = Disease
+        fields = ['name', 'ontology_terms']
+
+class GeneDiseaseSerializer(serializers.ModelSerializer):
+    disease = serializers.CharField()
+    identifier = serializers.CharField()
+    source = serializers.CharField(source="source.name")
+
+    class Meta:
+        model = GeneDisease
+        fields = ['disease', 'identifier', 'source']
